cursor_x value must be integer

+++ lib/eco/nodeeditor.py
@@ -493,7 +493,7 @@ class NodeEditor(QFrame):
             line += 1
 
         self.tm.cursor.line = line
-        cursor_x = round(float(x) / self.fontwt)
+        cursor_x = int(round(float(x) / self.fontwt))
         self.tm.cursor.move_to_x(cursor_x, self.tm.lines)
 
         if mouse_y > y or self.tm.cursor.get_x() != cursor_x:
